Create linear search
Linear searching techniques are the simplest technique. In this technique, the items are searched one by one. This procedure is also applicable for unsorted data set. Linear search is also known as sequential search. It is named as linear because its time complexity is of the order of n O(n).

The complexity of Linear Search Technique
Time Complexity: O(n)
Space Complexity: O(1)