Create Optimizing bubble sort
Bubble Sort is a comparison based sorting algorithm. In this algorithm adjacent elements are compared and swapped to make the correct sequence. This algorithm is simpler than other algorithms, but it has some drawbacks also. This algorithm is not suitable for a large number of data set. It takes much time to solve the sorting tasks.

The complexity of the Bubble Sort Technique
Time Complexity: O(n) for best case, O(n^2) for average and worst case
Space Complexity: O(1)